build: rename main => ip
