+++ lib/public_api.d.ts
@@ -1,1 +1,1 @@
-export * from './src/main';
+export * from './src/ip';
