+++ lib/src/ip.js
@@ -31,4 +31,4 @@ function getIpv6() {
     return IPv6;
 }
 exports.getIpv6 = getIpv6;
-//# sourceMappingURL=main.js.map
+//# sourceMappingURL=ip.js.map
